Updated Melee Attack to add hitstop to heavy attack
Also experimenting with adding ranged attack to player. Note Ranged Attack 01 Container has a start delay. Ranged Attack 02 does not. This start delay is so we can play a proper animation. The AI does not require a start delay, because there is a wait time set in the behavior designer. To fix mismatch issues, the behavior script for performing a ranged attack should just play the ranged attack and relay on the start delay time.