Adding multiple files from nuededad collaboration
The file modifications listed here are plentiful.Behavior designer files were modified to perform agnostic to whether there is a navAgent attached to the attached gameobject. Additional behaviors were added to create an AI Level Generator. MeleeAttacks were modified to aim in the direction of a reticule when user is in first person view or in aim view. This was done to adapt existing code to the Time Crisis style game The Demon at Dusk. UI scripts were added to include an FPS reticle, and a Fade in and out effect for text.