Added logic to allow for AI to self destruct on attack
Nuedeads game requires bazoop and others to self destruct causing massive damage. I added a variable to CollisionComponent to check if the attack should damage player. In the MeleeAttackController I added logic to deal damage to self based on a var set in collisionComponent. The health damage is done after the animation plays, this is so the explosion aspect of the move triggers before the enemy game object is destroyed