Modified and Added multiple scripts tohandle completing a game loop.
This diff includes code hastily written to meet a deadline. There are some good logical moves here, and some stumbles that can be learned from. The loop needed to be closed on transitioning between level general waves, summoning a boss, then winning or losing and proceeding to the appropriate screen afterwards.Logic here also includes a patch for detaching an overlay camera from its parent. This was eventually moved to its own script.